pyproject.toml
[project]
name = "sheriapal-backend"
version = "0.1.0"
description = "Legal document platform with AI integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.8",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.23.0",
    "uvicorn>=0.40.0",
    "argon2-cffi>=21.1",
    "slowapi>=0.1.9",
]

[tool.setuptools]
packages = ["app", "app.routers", "app.services"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.20",
    "coverage>=6.0",
]
